Refactor channel functionality into core
