+++ packages/feathers/src/http/utils.ts
@@ -94,46 +94,3 @@ export function getResponse(context: HookContext) {
 }
 
 export type QueryStringify = (query: Query) => string
-
-export function stringifyQuery(query: Query = {}) {
-  const searchParams = new URLSearchParams()
-
-  // Add each query parameter to URLSearchParams
-  Object.entries(query).forEach(([key, value]) => {
-    if (value !== undefined) {
-      if (Array.isArray(value)) {
-        // Handle array values
-        value.forEach((item) => searchParams.append(key, String(item)))
-      } else if (typeof value === 'object' && value !== null) {
-        // Handle object values by stringifying them
-        searchParams.append(key, JSON.stringify(value))
-      } else {
-        // Handle primitive values
-        searchParams.append(key, String(value))
-      }
-    }
-  })
-
-  return searchParams.toString()
-}
-
-export type QueryParser = (queryString: string) => Query
-
-export function parseQuery(queryString: string): Query {
-  const searchParams = new URLSearchParams(queryString)
-  const query: Query = {}
-
-  for (const [key, value] of searchParams.entries()) {
-    if (query[key] !== undefined) {
-      // Handle multiple values for the same key (arrays)
-      if (!Array.isArray(query[key])) {
-        query[key] = [query[key]]
-      }
-      ;(query[key] as any[]).push(value)
-    } else {
-      query[key] = value
-    }
-  }
-
-  return query
-}
